% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aresClass_Show.R
\docType{class}
\name{ARES-class}
\alias{ARES-class}
\title{Ares class representation: class definition and show method}
\description{
Ares class representation: class definition and show method
}
\keyword{internal}

