\name{Internal}
\alias{get.extension}
\alias{is.windows}
\alias{model.family}
\alias{perc.range}

\title{Internal function
}
\description{
Internal functions called from within \code{ares}
}
\details{
These functions are not intended to user-level calling.
}
\author{Washington Junger \email{wjunger@ims.uerj.br} and Antonio Ponce de Leon \email{ponce@ims.uerj.br}
}
\keyword{internal}
