% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.genepop.R
\name{read.genepop}
\alias{read.genepop}
\title{Imports a genepop file}
\usage{
read.genepop(filename = NULL)
}
\arguments{
\item{filename}{path (absolute or relative) & filename of genepop file to be read.}
}
\value{
The function returns an object that holds information regarding the genepop file.
The output can be used by the function \code{aresCalc}.
}
\description{
Imports a genepop formatted file.
}
\examples{
# import of genepop file butterfly_borneo
path <- system.file( package = 'ARES' )
file <- paste( list( path, '/extdata/butterfly_borneo' ), collapse="" )

butterfly_data <- read.genepop( filename=file )

#summary of the ARES object
butterfly_data

#rarefaction analysis for the fourth population
output <- aresCalc(butterfly_data,pop=4, bootsize=4, maxsize=100)
aresPlot(output, title='I2, 1998')

# note!!! this example uses a very small bootstrap size (for speed of display).
#  A more realistic size is bootsize=200.
}
\author{
Scott Davis, based on code by Emiel van Loon
}
\references{
Raymond M, Rousset F (1995) GENEPOP (version 1.2): population genetics software for exact tests and ecumenicism. J. Heredity, 86, 248-249.

\url{http://wbiomed.curtin.edu.au/genepop/help_input.html}
}

